Add new file menu and for files for creating book, person, rental, student and teacher for refactoring the code
