+++ create_a_book.rb
@@ -1,0 +1,13 @@
+class CreateBook
+  def create_a_book
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    book = Book.new(title, author)
+    @books << book
+    puts 'Book added successfully'
+    sleep 0.65
+    menu
+  end
+end
