+++ create_a_rental.rb
@@ -1,0 +1,21 @@
+class CreateRental
+  def create_a_rental
+    puts 'Select a book from the following list by number'
+    @books.each_with_index { |book, index| puts "#{index} Title: #{book.title}, Author: #{book.author}" }
+    book_id = gets.chomp.to_i
+
+    puts 'Select a person from the following list by number (not id)'
+    @people.each_with_index do |person, index|
+      puts "#{index} [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+    end
+    person_id = gets.chomp.to_i
+
+    print 'Date: '
+    date = gets.chomp.to_s
+    rental = Rental.new(date, @people[person_id], @books[book_id])
+    @rentals.push(rental)
+    puts 'Rental created successfully'
+    sleep 0.65
+    menu
+  end
+end
